Test for list_objects example
